Update form elements related storybooks
